requirements.txt
scikit-learn
flask
click
beautifulsoup4
pytest-cov